+++ src/components/ui/AdminComponents/ProvincesAdmin/ProvincesAdmin.tsx
@@ -1,4 +1,4 @@
-import { useEffect } from 'react'
+import { useEffect, useState } from 'react'
 import styles from './ProvincesAdmin.module.css'
 import { useStoreProvince } from '../../../../store/useStoreProvince'
 import { deleteProvince } from '../../../../cruds/crudProvince'
@@ -7,18 +7,32 @@ import { errorAlert } from '../../../../utils/errorAlert'
 import { useStoreModal } from '../../../../store/useStoreModal'
 import { AdminProvince } from '../../Modals/AdminProvince/AdminProvince'
 import { IProvince } from '../../../../types/IProvince'
+import { getAllLocalities } from '../../../../cruds/crudLocality'
+import { ILocality } from '../../../../types/ILocality'
 
 export const ProvincesAdmin = () => {
 
     const {fetchProvince, provinces, setActiveProvince} = useStoreProvince()
     const {modalAdminProvince, openModalAdminProvince} = useStoreModal()
+    const [localities, setLocalities] = useState<ILocality[]>()
 
     useEffect(() => {
+        const getEntities = async() => {
+            const localitiesFetched = await getAllLocalities()
+            setLocalities(localitiesFetched)
+        }
+        getEntities()
         fetchProvince()
     },[])
 
     const handleDelete = async(provinceId : string) => {
     
+        const existInLocality = localities?.find((locality) => locality.province.id === Number(provinceId))
+        if (existInLocality) { 
+            errorAlert('Error', 'Una localidad depende de esta provincia')
+            return
+        }
+
         try {
             await deleteProvince(provinceId)
             succesAlert('Eliminado', 'Se elimino correctamente la provincia')
